requirements.txt
# 基础依赖
streamlit
# Streamlit 服务器部署依赖
altair<5
protobuf==3.20.3

# LangChain生态系统
streamlit
langchain_deepseek
langchain_core
langchain
langchain_chroma
langchain_community
langchain_openai

# 向量数据库
chromadb
pysqlite3-binary

# AI模型API
dashscope

# 用户管理和数据库
sqlite3
hashlib
uuid
datetime
re

# 数据处理库
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.0.0

# 安全和加密
bcrypt
cryptography
